REFACTOR: Move `ResultList` to reusable components of results pages
