+++ src/pages/Results/pages/FootO/pages/Splits/FootOSplits.tsx
@@ -1,5 +1,5 @@
 
-import ResultListContainer from "../../components/ResultListContainer.tsx";
+import ResultListContainer from "../../../../components/ResultsList/ResultListContainer.tsx";
 import {Alert} from "@mui/material";
 import ScienceIcon from "@mui/icons-material/Science";
 import {useTranslation} from "react-i18next";
